+++ diesel_cli/src/database_error.rs
@@ -4,10 +4,10 @@ use std::convert::From;
 use std::{fmt, io};
 use std::error::Error;
 
-use self::SetupError::*;
+use self::DatabaseError::*;
 
 #[derive(Debug)]
-pub enum SetupError {
+pub enum DatabaseError {
     #[allow(dead_code)]
     CargoTomlNotFound,
     IoError(io::Error),
@@ -15,25 +15,25 @@ pub enum SetupError {
     ConnectionError(result::ConnectionError),
 }
 
-impl From<io::Error> for SetupError {
+impl From<io::Error> for DatabaseError {
     fn from(e: io::Error) -> Self {
         IoError(e)
     }
 }
 
-impl From<result::Error> for SetupError {
+impl From<result::Error> for DatabaseError {
     fn from(e: result::Error) -> Self {
         QueryError(e)
     }
 }
 
-impl From<result::ConnectionError> for SetupError {
+impl From<result::ConnectionError> for DatabaseError {
     fn from(e: result::ConnectionError) -> Self {
         ConnectionError(e)
     }
 }
 
-impl Error for SetupError {
+impl Error for DatabaseError {
     fn description(&self) -> &str {
         match *self {
             CargoTomlNotFound => "Unable to find Cargo.toml in this directory or any parent directories.",
@@ -44,13 +44,13 @@ impl Error for SetupError {
     }
 }
 
-impl fmt::Display for SetupError {
+impl fmt::Display for DatabaseError {
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
         self.description().fmt(f)
     }
 }
 
-impl PartialEq for SetupError {
+impl PartialEq for DatabaseError {
     fn eq(&self, other: &Self) -> bool {
         match (self, other) {
             (