+++ CHANGELOG.md
@@ -34,6 +34,9 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 * Added `sum` and `avg` functions.
 
+* Added the `diesel setup`, `diesel database setup`, and `diesel database
+  reset` commands to the CLI.
+
 ### Changed
 
 * Rename both the `#[derive(Queriable)]` attribute and the `Queriable` trait to